requirements.txt
numpy
matplotlib
pycryptodome